Moved unit tests for MoveNotesBetweenDecks() into CardRepositoryTests.
