+++ AnkiJapaneseFlashcardManagerTest/Anki2CardControllerTests.cs
@@ -16,36 +16,6 @@ namespace AnkiJapaneseFlashcardManagerTests
 		string _anki2FolderPath = "./Resources/Anki2 Files/";
 
 		[Theory]
-		[InlineData("飲newKanji_食欠人良resourceKanji_decks.anki2", new[] { 1707169497960, 1707169570657, 1707169983389, 1707170000793 }, 1707160682667)]
-		public void Move_notes_between_decks(string anki2File, long[] noteIdsToMove, long deckIdToMoveTo)
-		{
-			//Arrange
-			string originalInputFilePath = _anki2FolderPath + anki2File;
-			string tempInputFilePath = $"{_anki2FolderPath}temp_{Guid.NewGuid()}.anki2";
-			File.Copy(originalInputFilePath, tempInputFilePath, true);//Copy the input file to prevent changes between unit tests
-			Anki2Context dbContext = new Anki2Context(tempInputFilePath);
-			Anki2Controller anki2Controller = new Anki2Controller(dbContext);
-			List<Card> originalNoteDeckJunctions = anki2Controller.GetTable<Card>()
-																.Where(c => noteIdsToMove.Contains(c.NoteId))
-																.ToList();//Grab the current note/deck relations for the give note ids
-
-			//Act
-			bool movedNotes = anki2Controller.MoveNotesBetweenDecks(noteIdsToMove, deckIdToMoveTo);
-
-			//Assert
-			movedNotes.Should().BeTrue();//Function completed successfully
-			List<Card> finalNoteDeckJunctions = anki2Controller.GetTable<Card>()
-																.Where(c => noteIdsToMove.Contains(c.NoteId))
-																.ToList();//Grab the current note/deck relations for the give note ids after running the function
-			finalNoteDeckJunctions.Count().Should().Be(originalNoteDeckJunctions.Count());//No note/deck relations should have been removed/added
-			finalNoteDeckJunctions.Select(c => c.DeckId).Should().AllBeEquivalentTo(deckIdToMoveTo);//All junction deckIds should be the given deckId
-
-			//Cleanup
-			DbContextHelper.ClearSqlitePool(dbContext);
-			File.Delete(tempInputFilePath);
-		}
-
-		[Theory]
 		[InlineData("emptyLearningKanji_0ivl飲1ivl食欠良5ivl人newKanji_decks.anki2", new[] { 1707169522144, 1707169497960, 1707169570657, 1707170000793, 1707169983389 }, 1, new[] { 1707169497960, 1707169570657, 1707170000793, 1707169983389 })]
 		[InlineData("emptyLearningKanji_0ivl飲1ivl食欠良5ivl人newKanji_decks.anki2", new[] { 1707169522144, 1707169497960, 1707169570657, 1707170000793, 1707169983389 }, 4, new[] { 1707169983389 })]
 		[InlineData("emptyLearningKanji_0ivl飲1ivl食欠良5ivl人newKanji_decks.anki2", new[] { 1707169522144, 1707169497960, 1707169570657, 1707170000793, 1707169983389 }, 5, new[] { 1707169983389 })]
